code/python-sb-client/requirements.txt
azure-servicebus==0.21.1
